Refactor structure and add subversion job
