+++ rust/lib-dlock/Cargo.toml
@@ -1,19 +1,19 @@
 [package]
-name = "dlock"
+name = "libdlock"
 version = "0.1.0"
 edition = "2021"
+
+
+[features]
+default = ["combiner_stat"]
+combiner_stat = []
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
 bindgen = "0.68.1"
 
-[features]
-default = ["combiner_stat"]
-combiner_stat = []
-
-
-[dependencies]
+[dev-dependencies]
 clap = { version = "4.3.12", features = ["derive"] }
 strum = { version = "0.25", features = ["derive"] }
 csv = "1.2.1"
@@ -21,20 +21,24 @@ itertools = "0.11.0"
 serde = { version = "1", features = ["derive", "rc"] }
 serde_with = { version = "3.0.0" }
 serial_test = { version = "2.0.0" }
-libdlock = {path = "./lib-dlock"}
-core_affinity = "0.8.0"
-quanta = "0.11.1"
-criterion = { version = "0.5.1", features = ["html_reports"] }
-debug_print = "1.0.0"
-thread-priority = "0.13.1"
-coalesce = "0.1.1"
 
+
+[dependencies]
+core_affinity = "0.8.0"
+intrusive-collections = "0.9.5"
+linux-futex = "0.2.0"
+num_cpus = "1.15.0"
+once_cell = "1.17.1"
+quanta = "0.11.1"
+thread_local = "1.1"
+criterion = { version = "0.5.1", features = ["html_reports"] }
+enum_dispatch = "0.3.11"
+crossbeam = "0.8.2"
+crossbeam-skiplist = "0.1.1"
+proc-macro2 = { version = ">=1.0.63" }
+fastrand = "2.0.0"
 
 
 [profile.release-with-debug]
 inherits = "release"
 debug = true
-
-[[bench]]
-name = "benchmark"
-harness = false
