+++ rust/lib-dlock/src/ccsynch.rs
