+++ rust/lib-dlock/src/ccsynch/node.rs
