+++ rust/lib-dlock/src/mutex_extension.rs
